server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["MarteOued <marteoued11@gmail.com>"]
package-mode = false  # Désactive l'installation du package
packages = []  # Liste vide puisque nous désactivons le packaging

[tool.poetry.dependencies]
python = "^3.9"
django = "4.0"
django-prometheus = "^2.3.1"
celery = "^5.5.2"
psycopg2-binary = "^2.9.10"
django-cors-headers = "3.13.0"
djangorestframework = "3.13.1"
django-filter = "^23.5"  # Version installée

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
